+++ clang/docs/ReleaseNotes.rst
@@ -1132,8 +1132,8 @@ clang-format
   ``enum`` enumerator lists.
 - Add ``OneLineFormatOffRegex`` option for turning formatting off for one line.
 - Add ``SpaceAfterOperatorKeyword`` option.
-- Add ``FunctionLikeMacros`` option so that their invocations are skipped by
-  ``RemoveParentheses``.
+- Add ``MacrosSkippedByRemoveParentheses`` option so that their invocations are
+  skipped by ``RemoveParentheses``.
 
 clang-refactor
 --------------